feat(surveys): tabbed expanded row on survey responses
Generated-By: PostHog Code
Task-Id: fdef0980-2f44-490b-820b-91d57c0754a6
